Introduce layered CurrentCountryViewController
CurrentCountryViewController has a single responsibility, to fetch current
country.

When coordinator gets the country object from CurrentCountryVC it embeds
CountryDetailsVC into the CurrentCountryVC.

This way CountryDetailsVC has a responsibility to present data from
country object and CurrentCountryVC has a responsibility to get the
country object, thus CountryDetailsVC is reusable when the country
object is already available and therefore be reused in next commit -
presenting details when selecting a country from the list.